// react-js-ui/src/contexts/tweet/map/geo.d.ts
/*
 * Copyright 2021 Shingo OKAWA
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export interface Model {
  key: string,
  lon: number,
  lat: number,
  sum: number,
}

export interface Query {
  from: number,
  to: number,
  text: string,
  hashtags: string[],
}

export interface Response {
  cityCoord: {
    lon: number,
    lat: number,
  },
}

export interface Config {
  radius: number,
  blur: number,
  zoom: number,
}